+++ apps/validator/src/main/scala/com/daml/network/validator/automation/TopupMemberTrafficTrigger.scala
@@ -6,7 +6,6 @@ import com.daml.network.automation.{
   TaskSuccess,
   TriggerContext,
 }
-import com.daml.network.codegen.java.splice.amulet as amuletCodegen
 import com.daml.network.codegen.java.splice.wallet.install.amuletoperation.CO_BuyMemberTraffic
 import com.daml.network.codegen.java.splice.wallet.install.amuletoperationoutcome.{
   COO_BuyMemberTraffic,
@@ -22,11 +21,12 @@ import com.daml.network.environment.{
 }
 import com.daml.network.scan.admin.api.client.BftScanConnection
 import com.daml.network.store.MultiDomainAcsStore.QueryResult
-import com.daml.network.util.{AmuletConfigSchedule, CNNodeUtil, Contract, PrettyInstances}
-import com.daml.network.validator.config.BuyExtraTrafficConfig
+import com.daml.network.util.{AmuletConfigSchedule, Contract}
 import com.daml.network.validator.store.ValidatorStore
-import com.daml.network.validator.util.ExtraTrafficTopupParameters
-import com.daml.network.wallet.{UserWalletManager, UserWalletService}
+import com.daml.network.validator.util.ValidatorUtil
+import com.daml.network.wallet.util.{ExtraTrafficTopupParameters, TopupUtil, ValidatorTopupConfig}
+import com.daml.network.wallet.UserWalletManager
+import com.digitalasset.canton.config.NonNegativeFiniteDuration
 import com.digitalasset.canton.logging.pretty.{Pretty, PrettyPrinting}
 import com.digitalasset.canton.time.Clock
 import com.digitalasset.canton.topology.DomainId
@@ -45,7 +45,8 @@ class TopupMemberTrafficTrigger(
     store: ValidatorStore,
     connection: CNLedgerConnection,
     participantAdminConnection: ParticipantAdminConnection,
-    buyExtraTrafficConfig: BuyExtraTrafficConfig,
+    validatorTopupConfig: ValidatorTopupConfig,
+    grpcDeadline: Option[NonNegativeFiniteDuration],
     clock: Clock,
     walletManager: UserWalletManager,
     scanConnection: BftScanConnection,
@@ -67,9 +68,10 @@ class TopupMemberTrafficTrigger(
         .getConfigAsOf(clock.now)
         .decentralizedSynchronizer
       topupParameters = ExtraTrafficTopupParameters(
-        decentralizedSynchronizerConfig.fees,
-        buyExtraTrafficConfig,
-        context.config.pollingInterval,
+        validatorTopupConfig.targetThroughput,
+        validatorTopupConfig.minTopupInterval,
+        decentralizedSynchronizerConfig.fees.minTopupAmount,
+        validatorTopupConfig.topupTriggerPollingInterval,
       )
       _ = assert(topupParameters.topupAmount > 0, "topupAmount must be positive")
       activeSynchronizerId = DomainId.tryFromString(
@@ -79,19 +81,23 @@ class TopupMemberTrafficTrigger(
         activeSynchronizerId
       )
       topupState <- getOrCreateValidatorTopupState(activeSynchronizerId)
-      amuletPrice <- scanConnection.getLatestOpenMiningRound().map(_.payload.amuletPrice)
-      extraTrafficPrice = BigDecimal(decentralizedSynchronizerConfig.fees.extraTrafficPrice)
-      amuletBalance <- getWalletBalance()
+      validatorWallet <- ValidatorUtil.getValidatorWallet(store, walletManager)
+      hasSufficientFunds <- TopupUtil
+        .hasSufficientFundsForTopup(
+          scanConnection,
+          validatorWallet.store,
+          validatorTopupConfig,
+          clock,
+        )
     } yield {
       if (
+        // we do not even submit the topup tx if the validator does not have sufficient funds because we know
+        // the tx would fail but it would still drain synchronizer traffic which we would like to avoid (see #11915).
+        hasSufficientFunds &&
         shouldTopup(
           currentTrafficState,
           topupState,
           topupParameters,
-          amuletBalance,
-          amuletPrice,
-          extraTrafficPrice,
-          amuletRules.payload.isDevNet,
         )
       ) {
         Seq(
@@ -99,8 +105,6 @@ class TopupMemberTrafficTrigger(
             topupParameters,
             topupState,
             currentTrafficState,
-            amuletBalance,
-            amuletRules.payload.isDevNet,
           )
         )
       } else Seq.empty
@@ -119,8 +123,8 @@ class TopupMemberTrafficTrigger(
       Optional.of(task.topupState.contractId),
     )
     for {
-      validatorTreasury <- getValidatorWallet().map(_.treasury)
-      outcome <- validatorTreasury
+      validatorWallet <- ValidatorUtil.getValidatorWallet(store, walletManager)
+      outcome <- validatorWallet.treasury
         .enqueueAmuletOperation(coBuyMemberTraffic, CommandPriority.High)
         .map {
           case outcome: COO_BuyMemberTraffic =>
@@ -155,21 +159,11 @@ class TopupMemberTrafficTrigger(
       currentTrafficState: MemberTrafficStatus,
       topupState: Contract[ValidatorTopUpState.ContractId, ValidatorTopUpState],
       topupParameters: ExtraTrafficTopupParameters,
-      amuletBalance: BigDecimal,
-      amuletPrice: BigDecimal,
-      extraTrafficPrice: BigDecimal,
-      isDevNet: Boolean,
   )(implicit traceContext: TraceContext): Boolean = {
     val currentExtraTrafficRemainder = currentTrafficState.trafficState.extraTrafficRemainder.value
     val currentTime = clock.now
     val tooSoon =
       topupState.payload.lastPurchasedAt.toEpochMilli + topupParameters.minTopupInterval.duration.toMillis > currentTime.toEpochMilli
-    val (_, amuletsNeededForTopup) =
-      CNNodeUtil.synchronizerFees(
-        topupParameters.topupAmount,
-        extraTrafficPrice,
-        amuletPrice,
-      )
     if (tooSoon) {
       logger.trace(
         s"Trying to top-up too soon after previous top-up (last purchased at = ${topupState.payload.lastPurchasedAt}, current time = $currentTime)"
@@ -180,60 +174,8 @@ class TopupMemberTrafficTrigger(
         s"Sufficient traffic balance remains (current traffic balance = $currentExtraTrafficRemainder, topup amount = ${topupParameters.topupAmount})"
       )
       false
-    } else if (!isDevNet && amuletBalance < amuletsNeededForTopup) {
-      // We avoid submitting the top-up tx if there are insufficient funds even though this is enforced in daml
-      // because a failing tx will still consume traffic (see #11809).
-      // On DevNet (and in tests) we auto-tap coins which renders this check unnecessary. In this case, this check
-      // would actively slow down tests and make them flaky so we explicitly avoid doing it.
-      // TODO(#8046): Considering removing this once we remove auto-tapping in DevNet
-      logger.trace(
-        s"Insufficient funds to purchase traffic (current wallet balance = $amuletBalance, required = $amuletsNeededForTopup)"
-      )
-      false
     } else {
       true
-    }
-  }
-
-  private def getValidatorWallet()(implicit tc: TraceContext): Future[UserWalletService] = {
-    for {
-      walletInstall <- store
-        .lookupInstallByParty(store.key.validatorParty)
-        .map(
-          _.getOrElse(
-            throw Status.NOT_FOUND
-              .withDescription(
-                s"No wallet install contract for validator ${store.key.validatorParty}."
-              )
-              .asRuntimeException()
-          )
-        )
-      validatorWalletUser = walletInstall.payload.endUserName
-      validatorWallet = walletManager
-        .lookupUserWallet(validatorWalletUser)
-        .getOrElse(
-          throw Status.NOT_FOUND
-            .withDescription(
-              s"No wallet found for validator user $validatorWalletUser."
-            )
-            .asRuntimeException()
-        )
-    } yield validatorWallet
-  }
-
-  private def getWalletBalance()(implicit traceContext: TraceContext) = {
-    for {
-      validatorUserStore <- getValidatorWallet().map(_.store)
-      amulets <- validatorUserStore.multiDomainAcsStore.listContracts(
-        amuletCodegen.Amulet.COMPANION
-      )
-      currentRound <- scanConnection
-        .getLatestOpenMiningRound()
-        .map(_.payload.round.number)
-    } yield {
-      amulets.view
-        .map(c => BigDecimal(CNNodeUtil.currentAmount(c.payload, currentRound)))
-        .sum
     }
   }
 
@@ -261,7 +203,7 @@ class TopupMemberTrafficTrigger(
                 Instant.ofEpochSecond(0),
               ),
               priority = CommandPriority.High,
-              deadline = buyExtraTrafficConfig.grpcDeadline,
+              deadline = grpcDeadline,
             )
             .withDedup(
               CNLedgerConnection.CommandId(
@@ -291,19 +233,12 @@ object TopupMemberTrafficTrigger {
       topupParameters: ExtraTrafficTopupParameters,
       topupState: Contract[ValidatorTopUpState.ContractId, ValidatorTopUpState],
       trafficState: MemberTrafficStatus,
-      amuletBalance: BigDecimal,
-      isDevNet: Boolean,
   ) extends PrettyPrinting {
     override def pretty: Pretty[Task] =
       prettyOfClass[Task](
         param("topupParameters", _.topupParameters),
         param("topupState", _.topupState),
         param("trafficState", _.trafficState),
-        param(
-          "amuletBalance",
-          (x: Task) => x.amuletBalance.toString,
-          (x: Task) => !x.isDevNet,
-        )(PrettyInstances.prettyString),
       )
   }
 }