+++ apps/validator/src/main/scala/com/daml/network/validator/config/ValidatorAppConfig.scala
@@ -106,12 +106,12 @@ case class ValidatorDecentralizedSynchronizerConfig(
     url: Option[String] = None,
     buyExtraTraffic: BuyExtraTrafficConfig = BuyExtraTrafficConfig(),
 
-    /** amount of extra traffic reserved for transactions required to do traffic topups
+    /** amount of extra traffic reserved for high-priority transactions eg. topups
       *
       * Note that this will be ignored if the validator is not configured to do topups
-      * i.e. the target throughput is set to zero (its default value). See: trafficReservedForTopupsO.
+      * i.e. the target throughput is set to zero (its default value). See: reservedTrafficO.
       */
-    trafficReservedForTopups: NonNegativeLong = NonNegativeNumeric.tryCreate(200_000L),
+    reservedTraffic: NonNegativeLong = NonNegativeNumeric.tryCreate(200_000L),
 
     /** The validator's ledger client compares its remaining traffic balance against the reserved amount
       * on every command submission. This setting controls how long the traffic balance is cached before
@@ -121,11 +121,11 @@ case class ValidatorDecentralizedSynchronizerConfig(
       NonNegativeFiniteDuration.ofSeconds(1),
 ) {
 
-  /** Converts the trafficReservedForTopups into an Option that is set to None if the validator is not
+  /** Converts the reservedTraffic into an Option that is set to None if the validator is not
     * configured to do top-ups in the first place
     */
-  lazy val trafficReservedForTopupsO: Option[NonNegativeLong] =
-    if (buyExtraTraffic.targetThroughput.value <= 0L) None else Some(trafficReservedForTopups)
+  lazy val reservedTrafficO: Option[NonNegativeLong] =
+    if (buyExtraTraffic.targetThroughput.value <= 0L) None else Some(reservedTraffic)
 }
 
 // Validators are responsible for establishing connections to domains and so need more information than just a `SynchronizerConfig`